fix: resolve all mypy type errors (115->0) and ruff lint errors (2->0)
- Add type annotations to empty container literals across 19 files
- Add isinstance checks on _get_cached() returns to narrow Any types
- Import Sequence from collections.abc (ruff UP035)
- Add type: ignore comments for unavoidable Any operations
- Cast returns with float()/str()/bool() for untyped library calls
- Change finnhub_client._get to return dict|None with separate _get_list
- Fix implicit optional parameters (no_implicit_optional=true)
- Remove unused sys import in bench_advisor.py (ruff F401)

CI: mypy src/ passes with 0 errors, ruff check passes clean

+++ src/services/value_scanner.py
@@ -213,14 +213,14 @@ def _compute_margin_of_safety(d: dict) -> Optional[float]:
         return None
 
     mos = round((intrinsic - price) / intrinsic * 100, 1)
-    return mos
+    return float(mos)
 
 
 def _compute_fcf_yield(d: dict) -> Optional[float]:
     fcf = d.get("free_cash_flow")
     mcap = d.get("market_cap", 0)
     if fcf is not None and mcap > 0:
-        return round((fcf / mcap) * 100, 2)
+        return float(round((fcf / mcap) * 100, 2))
     return None
